Tutor toggle works now

+++ src/Rayku/ApiBundle/Controller/TutorController.php
@@ -7,6 +7,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use FOS\RestBundle\Controller\Annotations\View;
 use Nelmio\ApiDocBundle\Annotation\ApiDoc;
+use JMS\SecurityExtraBundle\Annotation\SecureParam;
 
 use Rayku\ApiBundle\Entity\Tutor;
 use Rayku\ApiBundle\Form\TutorType;
@@ -15,28 +16,55 @@ use Rayku\ApiBundle\Form\TutorType;
  */
 class TutorController extends Controller
 {
+	
 	/**
 	 * @View()
 	 * @ApiDoc(
+	 *   statusCodes={
+     *     200="Returned when successful",
+     *     403="Returned when the user is not authorized"
+     *   },
 	 *   description="Delete a tutor account"
 	 * )
+	 * @todo setup / configure ACL
 	 */
 	public function deleteTutorsAction(Tutor $entity)
 	{
+		if($entity->getUser() != $this->getUser()){
+			throw new AccessDeniedException();
+		}
 		$em = $this->getDoctrine()->getManager();
 		$em->remove($entity);
 		$em->flush();
 		 
 		return new Response(
-				json_encode(array('success' => true)),
-				200,
-				array('Content-Type'=>'application/json')
+			json_encode(array('success' => true)),
+			200,
+			array('Content-Type'=>'application/json')
 		);
 	}
 	
 	/**
 	 * @View()
 	 * @ApiDoc(
+	 *   statusCodes={
+     *     200="Returned when successful"
+     *   },
+	 *   description="Get a tutor record",
+	 *   output="Rayku\ApiBundle\Entity\Tutor"
+	 * )
+	 */
+	public function getTutorAction(Tutor $entity)
+	{
+		return $entity;
+	}
+	
+	/**
+	 * @View()
+	 * @ApiDoc(
+	 *   statusCodes={
+     *     200="Returned when successful"
+     *   },
 	 *   description="Get online tutors"
 	 * )
 	 */
@@ -44,18 +72,90 @@ class TutorController extends Controller
 	{
 		$em = $this->getDoctrine()->getManager();
 		$entities = $em->getRepository('RaykuApiBundle:Tutor')->findOnlineTutors(Tutor::expire_online);
-		return array('entities' => $entities);
+		return $entities;
 	}
 	
 	/**
 	 * @View()
 	 * @ApiDoc(
+	 *   statusCodes={
+     *     200="Returned when successful",
+     *     400="Returned when there is an error"
+     *   },
 	 *   description="Create/update tutor object",
 	 *   resource=true,
 	 *   input="Rayku\ApiBundle\Form\TutorType"
 	 * )
 	 */
 	public function postTutorsAction()
+	{
+		$entity = $this->getTutor();
+		 
+		// Since we cleared doctrine above need to get a new user entity from the DB
+		$em = $this->getDoctrine()->getManager();
+		$user = $em->getRepository('RaykuApiBundle:User')->find($this->getUser()->getId());
+		$entity->setUser($user);
+		return $this->processForm($entity);
+	}
+	
+	/**
+	 * @todo http://symfony.com/doc/2.1/cookbook/form/dynamic_form_modification.html#dynamic-generation-for-submitted-forms
+	 * @View()
+	 * @ApiDoc(
+	 *   description="Patch a tutor record",
+	 *   input="Rayku\ApiBundle\Form\TutorType"
+	 * )
+	 */
+	public function patchTutorsAction()
+	{
+		$entity = $this->getTutor();
+		
+		if(null === $entity->getId() && $entity->getUser() != $this->getUser()){
+			throw new AccessDeniedException();
+		}
+		
+		$valid = false;
+		foreach($this->get('request')->request->all() as $k => $v){
+			if(in_array($k, array('gtalk_email'))){
+				$method = 'set' . preg_replace('/(?:^|_)(.?)/e',"strtoupper('$1')",$k);
+				if(method_exists($entity, $method)){
+					$entity->$method($v);
+					$valid = true;
+				}
+			}
+		}
+		
+		if(!$valid){
+			return array(
+				'success' => false, 
+				'message' => 'Invalid parameters'
+			);
+		}
+		
+		$errors = $this->get('validator')->validate($entity);
+		
+		if(count($errors) > 0){
+			$errorOutput = array();
+			foreach($errors as $error){
+				$errorOutput[$error->getPropertyPath()] = $error->getMessage();
+			}
+			return array(
+				'success' => false,
+				'form' => $errorOutput	
+			);
+		}
+		
+		$em = $this->getDoctrine()->getManager();
+		$em->persist($entity);
+		$em->flush();
+		
+		return array(
+			'success' => true,
+			'entity' => $entity	
+		);
+	}
+	
+	private function getTutor()
 	{
 		$em = $this->getDoctrine()->getManager();
 		$em->getFilters()->disable('soft_deleteable');
@@ -67,11 +167,7 @@ class TutorController extends Controller
 		}else{
 			$entity->setDeletedAt(NULL);
 		}
-		 
-		// Since we cleared doctrine above need to get a new user entity from the DB
-		$user = $em->getRepository('RaykuApiBundle:User')->find($this->getUser()->getId());
-		$entity->setUser($user);
-		return $this->processForm($entity);
+		return $entity;
 	}
 	
 	private function processForm(Tutor $entity)
@@ -109,8 +205,8 @@ class TutorController extends Controller
 		}
 		 
 		return array(
-				'entity' => $entity,
-				'form'   => $form,
+			'entity' => $entity,
+			'form'   => $form,
 		);
 	}
 }
